requirements.txt
python-dotenv==1.0.0
openai==0.28.0
pytest==7.4.3
streamlit==1.32.0 